// models/CollectionModelsV2/MonthsAvailableModel.ts
export interface MonthsAvaliableModel {
    jan: boolean | undefined;
    feb: boolean | undefined;
    mar: boolean | undefined;
    apr: boolean | undefined;
    may: boolean | undefined;
    jun: boolean | undefined;
    jul: boolean | undefined;
    aug: boolean | undefined;
    sep: boolean | undefined;
    oct: boolean | undefined;
    nov: boolean | undefined;
    dec: boolean | undefined;
}